+++ CHANGELOG.rst
@@ -32,6 +32,7 @@ number of the code change for that issue.  These PRs can be viewed at:
   more helpful. The request specification is also sent to the log, so the user
   can see what was actually requested. [#1451]
 
+
 3.5.0 (31-Aug-2022)
 ====================